requirements.txt
pandas==2.2.0
matplotlib==3.8.2
plotly==5.18.0
geopandas==0.14.2
pydeck==0.8.1b0
streamlit==1.31.1
imageio==2.34.0
numpy==1.26.0
scipy==1.11.3
jinja2==3.1.2
boto3==1.28.0
reverse_geocoder==1.5.1
rasterio==1.3.8
shapely==2.0.1
turfpy==0.0.7
openquake.hazardlib==3.12.0
requests==2.31.0
tqdm==4.66.1
statsmodels==0.14.0